% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_functions.R
\name{calc_mode}
\alias{calc_mode}
\title{Calculate mode}
\usage{
calc_mode(v)
}
\arguments{
\item{v}{Vector of numerical values}
}
\value{
number- mode of the supplied vector
}
\description{
Funnily enough, R doesn't have a mode, so this is it. 
Be careful because ties just return whichever is the furthest to the 
beginning of the vector with no warning given
}
