% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_and_forecast.R
\name{forecast_series}
\alias{forecast_series}
\title{Make a forecast of drug issues}
\usage{
forecast_series(data, horizon, frequency = "Daily")
}
\arguments{
\item{data}{tsibble with two columns 'Date' and 'Quantity'. You can make 
this with \code{\link{make_tsibble}}}

\item{horizon}{integer. number of days/ weeks to forecast}

\item{frequency}{String. One of 'Daily' or 'Weekly'- return daily or 
weekly forecasts}
}
\value{
mable with different forecast functions applied - mean, 
snaive (weekly seasonality), ARIMA (auto), and ETS (weekly seasonality for 
daily forecast, and none with weekly forecast)
}
\description{
Make a forecast of drug issues
}
