% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_functions.R
\name{get_holidays}
\alias{get_holidays}
\title{Get holidays}
\usage{
get_holidays()
}
\value{
vector of dates as string YYYY-MM-DD
}
\description{
Download English bank and other holidays from gov.uk
}
