% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_reorder.R
\name{inventory_reorder}
\alias{inventory_reorder}
\title{Forecast and run inventory management algorithm, returning reorder quantities
by site and drug}
\usage{
inventory_reorder(site, supplier)
}
\arguments{
\item{site}{String. Site code (selected from dynamic UI listing all sites)}

\item{supplier}{String. Supplier (changes weekly, selected in Shiny interface)}
}
\value{

}
\description{
Forecast and run inventory management algorithm, returning reorder quantities
by site and drug
}
