% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_quantiles.R
\name{make_quantiles}
\alias{make_quantiles}
\title{Make quantiles from a forecast}
\usage{
make_quantiles(forecast, model = "ETS", num_q_vals = 10)
}
\arguments{
\item{forecast}{dataframe with two columns- mean_demand and sd_demand}

\item{num_q_vals}{number of quantiles plus 1 (defaults to 10)}
}
\value{
matrix of quantiles
}
\description{
Make quantiles from a forecast
}
