% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_and_forecast.R
\name{make_tsibble}
\alias{make_tsibble}
\title{Make a tsibble suitable for forecsating}
\usage{
make_tsibble(data, frequency = "Daily", remove_weekends = FALSE)
}
\arguments{
\item{data}{Dataframe with two columns 'Date', and 'Total_Qty'}

\item{frequency}{String. One of 'Daily' or 'Weekly'}

\item{remove_weekends}{Boolean. Set to TRUE to remove weekends}
}
\value{
tsibble with two columns 'Date' and 'quantity'
}
\description{
Make a tsibble suitable for forecsating
}
