% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_functions.R
\name{n_weekdays}
\alias{n_weekdays}
\title{return number of weekdays}
\usage{
n_weekdays(from, to, holidays)
}
\arguments{
\item{from}{Date. Date to count from (inclusive)}

\item{to}{Date. Date to count to (inclusive)}

\item{holidays}{you can get this (for England) with 
\code{\link{get_holidays}}}
}
\value{
integer. Number of weekdays between two dates
}
\description{
return the number of weekdays between two days, excluding 
weekends (obviously) and holidays (user defined)
}
