% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_and_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot the forecast}
\usage{
plot_forecast(forecast_value, data, horizon, model)
}
\arguments{
\item{forecast_value}{A forecast. You can make this with 
\code{\link{forecast_series}}}

\item{data}{A tsibble, the same one the forecast is based on}
}
\value{
A faceted ggplot
}
\description{
Produce a faceted plot showing the performance of different 
forecasting models compared with the actual issues over that time span
}
