% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_and_forecast.R
\name{show_accuracy}
\alias{show_accuracy}
\title{Show the accuracy of a set of forecasts}
\usage{
show_accuracy(forecast_value, data, horizon)
}
\arguments{
\item{data}{tsibble with two columns 'Date' and 'Quantity'. You can make 
this with \code{\link{make_tsibble}}}

\item{forecast_value.}{A forecast. You can make this with 
\code{\link{forecast_series}}}
}
\value{
dataframe with two columns showing model and MAE
}
\description{
Take a mable of forecasts and the actual drug issues and compute
mean absolute error
}
